requirements.txt
django>=1.8.4
django-braces
